requirements.txt
# This file was autogenerated by uv via the following command:
#    uv pip compile - -o requirements.txt
annotated-types==0.7.0
    # via pydantic
anyio==4.4.0
    # via httpx
certifi==2024.7.4
    # via
    #   httpcore
    #   httpx
exceptiongroup==1.2.2
    # via anyio
h11==0.14.0
    # via httpcore
httpcore==1.0.5
    # via httpx
httpx==0.27.0
    # via mistralai
idna==3.7
    # via
    #   anyio
    #   httpx
mistralai==0.4.2
neo4j==5.23.1
numpy==2.0.1
    # via pandas
orjson==3.10.6
    # via mistralai
pandas==2.2.2
pydantic==2.8.2
    # via mistralai
pydantic-core==2.20.1
    # via pydantic
python-dateutil==2.9.0.post0
    # via pandas
python-dotenv==1.0.1
pytz==2024.1
    # via
    #   neo4j
    #   pandas
six==1.16.0
    # via python-dateutil
sniffio==1.3.1
    # via
    #   anyio
    #   httpx
typing-extensions==4.12.2
    # via
    #   anyio
    #   pydantic
    #   pydantic-core
tzdata==2024.1
    # via pandas
